# Conflicts:
#	build.gradle
#	grails-app/controllers/au/org/ala/spatial/service/LoginInterceptor.groovy
#	grails-app/controllers/au/org/ala/spatial/service/MasterController.groovy
#	grails-app/services/au/org/ala/spatial/service/MasterService.groovy
#	grails-app/services/au/org/ala/spatial/slave/SlaveService.groovy
#	src/main/groovy/au/org/ala/spatial/process/AooEoo.groovy
#	src/main/groovy/au/org/ala/spatial/process/DistributionRematchLsid.groovy
#	src/main/groovy/au/org/ala/spatial/process/GeneratePoints.groovy
#	src/main/groovy/au/org/ala/spatial/process/PhylogeneticDiversity.groovy
#	src/test/resources/resources/output/aooeoo/Calculated AOO and EOO.html

+++ src/main/groovy/au/org/ala/spatial/process/AooEoo.groovy
@@ -46,7 +46,7 @@ class AooEoo extends SlaveProcess {
         def species = JSON.parse(taskWrapper.input.species.toString())
 
         // concave hull coverage parameter
-        def alpha = taskWrapper.input.coverage
+        def alpha = (Double) taskWrapper.input.coverage
 
         def speciesArea = getSpeciesArea(species, area[0])
